requirements.txt
bcrypt==4.1.3
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.0.0
Flask-WTF==1.2.2
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.2
schedule==1.2.0
urllib3==2.5.0
Werkzeug==3.0.3
WTForms==3.2.1
